Admin posts and comments management

+++ Project/LoveIT/src/main/java/com/group03/loveit/controllers/favorites/RemoveFavoriteController.java
@@ -2,6 +2,8 @@ package com.group03.loveit.controllers.favorites;
 
 import com.group03.loveit.controllers.post_details.CreateCommentController;
 import com.group03.loveit.models.favourite.FavoriteDAO;
+import com.group03.loveit.models.user.UserDTO;
+import com.group03.loveit.utilities.ConstantUtils;
 
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
@@ -13,7 +15,14 @@ public class RemoveFavoriteController extends HttpServlet {
     protected void doPost(HttpServletRequest request, HttpServletResponse response) throws IOException {
         try {
             long postId = Long.parseLong(request.getParameter("post_id"));
-            long userId = CreateCommentController.getCurrentUser().getId();
+            UserDTO user = (UserDTO) request.getSession().getAttribute(ConstantUtils.SESSION_USER);
+
+            if (user == null) {
+                response.sendRedirect(request.getContextPath() + "/login");
+                return;
+            }
+
+            long userId = user.getId();
 
             FavoriteDAO favoriteDAO = new FavoriteDAO();
             favoriteDAO.deleteFavorite(postId, userId).join();
